requirements.txt
python-dotenv
llama-index
nest_asyncio
ipykernel